requirements.txt
streamlit>=1.35.0
pinecone>=3.0.0
openai>=1.30.0
httpx>=0.27.0
beautifulsoup4>=4.12.0
pdfplumber>=0.11.0
python-docx>=1.1.0
